Add support for signed `.pkg` archives on macOS
