+++ Sources/CartonHelpers/DefaultToolchain.swift
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-public let defaultToolchainVersion = "wasm-5.3-SNAPSHOT-2020-10-15-a"
+public let defaultToolchainVersion = "wasm-5.3-SNAPSHOT-2020-10-20-a"
